fix: inline import persistence in ImportService
Moved import persistence logic from `utils/persist.ts` into `ImportService` as private methods (`toMessageDto` and `persistImport`), and updated imports/types accordingly. This keeps parsing and DataApi persistence in one place while preserving the existing topic/message creation flow and model snapshot mapping. Removed the now-redundant `persist.ts` helper file.
